+++ Changelog.md
@@ -3,6 +3,7 @@
 - New: Block Change Detector. Detects and logs players breaking and/or placing blocks in its vicinity
 - New: Codebreakers can now hack Keycard Readers
 - Change: Codebreakers will now always open briefcases when in Creative mode without a chance of failing
+- Change: The Sentry's body is now solid
 - API: Split IPasswordProtected into two interfaces (IPasswordProtected and ICodebreakable), meaning blocks can now be hacked by the codebreaker without needing to be password protected
 - API: Removed IPasswordProtected#isCodebreakable
 - Fix: The recipe for reinforced glass panes is incorrect
@@ -11,6 +12,7 @@
 - Fix: Some logs have an incorrect side texture
 - Fix: Reinforced Iron Trapdoor textures do not rotate and aren't oriented the same way open as closed
 - Fix: Reinforced Ice blocks melt unintentionally
+- Fix: Disguised blocks from sentries can be retrieved by using pistons or other means
 - Removed: Some unnecessary menu types. This may result in a "missing registry entries" message showing up, which can be accepted
 - Misc.: More texture and model changes